(#72) Change some package hierarchy
