+++ src/main/java/org/fusionsoft/database/mapping/package-info.java
@@ -13,27 +13,8 @@
  * See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package org.fusionsoft.database.snapshot.databaseinfo;
-
-import org.cactoos.Text;
-import org.fusionsoft.database.snapshot.databaseinfo.dbmssignature.DbmsSignatureName;
 
 /**
- * The interface DbmsSignature representing DBMS's name and version.
- * @since 0.1
+ * Various yaml mappings artefacts.
  */
-public interface DbmsSignature {
-
-    /**
-     * Kind of DBMS used e.g. Postgres.
-     * @return The Text.
-     */
-    DbmsSignatureName name();
-
-    /**
-     * Version of DBMS used.
-     * @return The Text.
-     */
-    Text version();
-
-}
+package org.fusionsoft.database.mapping;
