+++ src/main/java/org/fusionsoft/database/writable/package-info.java
@@ -13,24 +13,8 @@
  * See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package org.fusionsoft.database;
-
-import com.amihaiemil.eoyaml.Yaml;
-import org.fusionsoft.lib.yaml.YamlMappingOf;
 
 /**
- * The type of YamlMapping which is just empty.
- * @since 0.1
+ * {@link org.fusionsoft.database.Writable} implementations.
  */
-public class MappingEmpty extends YamlMappingOf {
-
-    /**
-     * Instantiates a new Mapping empty.
-     */
-    public MappingEmpty() {
-        super(
-            () -> Yaml.createYamlMappingBuilder().build()
-        );
-    }
-
-}
+package org.fusionsoft.database.writable;
